+++ src/features/scraps/mutations/useUnscrapOutfitItem.ts
@@ -1,7 +1,7 @@
 import { useMutation, useQueryClient } from '@tanstack/react-query';
 import { HttpStatusCode } from 'axios';
-import { unscrapOutfitItem } from '../../api/scrap';
-import { getServerErrorResponse } from '../../utils/error';
+import { unscrapOutfitItem } from '../../../api/scrap';
+import { getServerErrorResponse } from '../../../utils/error';
 
 const useUnscrapOutfitItem = () => {
   const queryClient = useQueryClient();
