+++ src/pages/MyScrapPage/components/MyScrapOutfitItemGrid.tsx
@@ -4,7 +4,7 @@ import Spinner from '../../../components/atoms/Spinner';
 import InfiniteScrollFetchTrigger from '../../../components/common/InfiniteScrollFetchTrigger';
 import Grid from '../../../components/templates/Grid';
 import OutfitScrapButton from '../../../features/outfits/components/OutfitScrapButton';
-import useFetchMyScrapOutfitItemList from '../../../hooks/queries/useFetchMyScrapOutfitItemList';
+import useFetchMyScrapOutfitItemList from '../../../features/scraps/queries/useFetchMyScrapOutfitItemList';
 import { MyScrapOutfitItemListResponse } from '../../../types/scrap';
 
 interface GridContentProps {
