+++ src/pages/MyScrapPage/components/MyScrapOutfitPostGrid.tsx
@@ -4,7 +4,7 @@ import Spinner from '../../../components/atoms/Spinner';
 import InfiniteScrollFetchTrigger from '../../../components/common/InfiniteScrollFetchTrigger';
 import Grid from '../../../components/templates/Grid';
 import OutfitScrapButton from '../../../features/outfits/components/OutfitScrapButton';
-import useFetchMyScrapOutfitPostList from '../../../hooks/queries/useFetchMyScrapOutfitPostList';
+import useFetchMyScrapOutfitPostList from '../../../features/scraps/queries/useFetchMyScrapOutfitPostList';
 import { MyScrapOutfitPostListResponse } from '../../../types/scrap';
 
 interface GridContentProps {
